Progress save

+++ controllers/userController.js
@@ -1,5 +1,6 @@
+const mongoose = require("mongoose");
 const { ObjectId } = require("mongoose").Types;
-const { User } = require("../models");
+const { User, Thought } = require("../models");
 
 module.exports = {
   // Get all users
@@ -20,6 +21,12 @@ module.exports = {
   // Get a single user
   async getSingleUser(req, res) {
     try {
+      const userId = req.params.userId;
+
+      if (!mongoose.isValidObjectId(userId)) {
+        return res.status(404).json({ message: "Invalid user ID" });
+      }
+
       const user = await User.findOne({
         _id: req.params.userId,
       }).select("-__v");
@@ -45,9 +52,38 @@ module.exports = {
       res.status(500).json(err);
     }
   },
+  // Update a user record
+  async updateUser(req, res) {
+    try {
+      const userId = req.params.userId;
+
+      if (!mongoose.isValidObjectId(userId)) {
+        return res.status(404).json({ message: "Invalid user ID" });
+      }
+
+      const user = await User.findOneAndUpdate(
+        { _id: req.params.userId },
+        { $set: req.body },
+        { new: true }
+      );
+
+      if (!user) {
+        return res.status(404).json({ message: "No such user exists" });
+      }
+    } catch (err) {
+      console.log(err);
+      res.status(500).json(err);
+    }
+  },
   // Delete a user and remove their associated thoughts
   async deleteUser(req, res) {
     try {
+      const userId = req.params.userId;
+
+      if (!mongoose.isValidObjectId(userId)) {
+        return res.status(404).json({ message: "Invalid user ID" });
+      }
+
       const user = await User.findOneAndRemove({
         _id: req.params.userId,
       });
@@ -56,19 +92,18 @@ module.exports = {
         return res.status(404).json({ message: "No such user exists" });
       }
 
-      const thoughts = await Thought.findAllAndUpdate(
-        { users: req.params.userId },
-        { $pull: { users: req.params.userId } },
-        { new: true }
-      );
-
-      if (!thoughts) {
+      if (!user.thoughts.length) {
         return res.status(404).json({
-          message: "Student deleted, but no thoughts found",
+          message: "User deleted, but no thoughts found",
         });
       }
 
-      res.json({ message: "Student successfully deleted" });
+      // Remove all thoughts associated with the deleted user
+      await Thought.deleteMany({ username: user.username });
+
+      res.json({
+        message: "User and associated thoughts successfully deleted",
+      });
     } catch (err) {
       console.log(err);
       res.status(500).json(err);
